moved useClickOutside to shared/hooks
